requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
albumentations==1.3.1
annotated-types==0.5.0
anyio==3.7.1
arrow==1.2.3
async-timeout==4.0.3
attrs==23.1.0
backoff==2.2.1
beautifulsoup4==4.12.2
blessed==1.20.0
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
croniter==1.4.1
datasets==2.14.5
dateutils==0.6.12
deepdiff==6.5.0
dill==0.3.7
fastapi==0.103.1
filelock==3.12.4
frozenlist==1.4.0
fsspec==2023.6.0
google-api-core==2.11.1
google-auth==2.23.0
google-cloud-texttospeech==2.14.1
googleapis-common-protos==1.60.0
grpcio==1.58.0
grpcio-status==1.58.0
h11==0.14.0
httptools==0.6.0
huggingface-hub==0.17.1
idna==3.4
imageio==2.31.3
inquirer==3.1.3
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
lazy_loader==0.3
Levenshtein==0.21.1
lightning==2.0.9
lightning-cloud==0.5.38
lightning-utilities==0.9.0
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.15
munch==4.0.0
networkx==3.1
nltk==3.8.1
nougat==0.3.3
nougat-ocr==0.1.8
numpy==1.25.2
opencv-python-headless==4.8.0.76
ordered-set==4.1.0
orjson==3.9.7
packaging==23.1
pandas==2.1.0
Pillow==10.0.1
proto-plus==1.22.3
protobuf==4.24.3
psutil==5.9.5
pyarrow==13.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic==2.1.1
pydantic_core==2.4.0
pydub==0.25.1
Pygments==2.16.1
PyJWT==2.8.0
PyMuPDF==1.23.3
PyMuPDFb==1.23.3
python-dateutil==2.8.2
python-editor==1.0.4
python-Levenshtein==0.21.1
python-multipart==0.0.6
pytorch-lightning==2.0.9
pytz==2023.3.post1
PyWavelets==1.4.1
PyYAML==6.0.1
qudida==0.0.4
rapidfuzz==3.3.0
readchar==4.0.5
regex==2023.8.8
requests==2.31.0
rich==13.5.2
rsa==4.9
ruamel.yaml==0.17.32
ruamel.yaml.clib==0.2.7
safetensors==0.3.3
scikit-image==0.21.0
scikit-learn==1.3.0
scipy==1.11.2
sconf==0.2.5
sentencepiece==0.1.99
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
starlette==0.27.0
starsessions==1.3.0
sympy==1.12
threadpoolctl==3.2.0
tifffile==2023.8.30
timm==0.5.4
tokenizers==0.13.3
torch==2.0.1
torchmetrics==1.1.2
torchvision==0.15.2
tqdm==4.66.1
traitlets==5.10.0
transformers==4.33.1
typing_extensions==4.7.1
tzdata==2023.3
urllib3==1.26.16
uvicorn==0.23.2
wcwidth==0.2.6
websocket-client==1.6.3
websockets==11.0.3
xxhash==3.3.0
yarl==1.9.2
